Vulkan clear screan

+++ voxl-core/src/app/game.cpp
@@ -6,24 +6,24 @@
 #include "math/vec4.hpp"
 
 namespace voxl {
-namespace app {
-Game::Game(Config config) : renderer(config) {
-  renderer.Submit(
-      new graphics::commands::ClearCommand(math::Vec4(0.3f, 0.3f, 0.3f, 1.0f)));
-}
+	namespace app {
+		Game::Game(Config config) : renderer(config) {
+			renderer.Submit(
+				new graphics::commands::ClearCommand(math::Vec4(0.3f, 0.3f, 0.3f, 1.0f)));
+		}
 
-void Game::start() {
-  float lastTime = glfwGetTime();
+		void Game::start() {
+			double lastTime = glfwGetTime();
 
-  while (!glfwWindowShouldClose(renderer.getWindow())) {
-    float curTime = glfwGetTime();
-    float delta = curTime - lastTime;
-    glfwPollEvents();
-    lastTime = curTime;
+			while (!glfwWindowShouldClose(renderer.getWindow())) {
+				double curTime = glfwGetTime();
+				double delta = curTime - lastTime;
+				glfwPollEvents();
+				lastTime = curTime;
 
-    renderer.Render();
-    // std::cout << "FPS: " << 1000.0f / (delta * 1000.0f) << std::endl;
-  }
-}
-}
+				renderer.Render();
+				// std::cout << "FPS: " << 1000.0f / (delta * 1000.0f) << std::endl;
+			}
+		}
+	}
 }
